requirements.txt
# Core dependencies
tqdm>=4.66.0
Pillow>=10.0.0

# Document processing (optional - install as needed)
python-docx>=1.0.0
PyMuPDF>=1.23.0
python-pptx>=0.6.21
openpyxl>=3.1.0

# API clients (optional - choose one)
anthropic>=0.18.0
openai>=1.0.0
